+++ mex/editor/search/models.py
@@ -18,3 +18,17 @@ class SearchPrimarySource(rx.Base):
     identifier: str
     title: str
     checked: bool
+
+
+class ReferenceFieldIdentifierFilter(rx.Base):
+    """Reference field identifier for value and validation msg."""
+
+    value: str
+    validation_msg: str | None
+
+
+class ReferenceFieldFilter(rx.Base):
+    """Reference field filter."""
+
+    identifiers: list[ReferenceFieldIdentifierFilter]
+    field: str