+++ mex/editor/search/state.py
@@ -1,24 +1,58 @@
 import math
 from collections.abc import Generator
-from typing import TYPE_CHECKING
+from typing import TYPE_CHECKING, Literal
 
 import reflex as rx
+from pydantic import TypeAdapter, ValidationError
 from reflex.event import EventSpec
 from requests import HTTPError
 
 from mex.common.backend_api.connector import BackendApiConnector
 from mex.common.exceptions import MExError
+from mex.common.fields import REFERENCE_FIELDS_BY_CLASS_NAME
 from mex.common.models import MERGED_MODEL_CLASSES, MergedPrimarySource
 from mex.common.transform import ensure_prefix
-from mex.editor.constants import DEFAULT_FETCH_LIMIT
+from mex.common.types import Identifier
+from mex.editor.constants import DEFAULT_FETCH_LIMIT, PRIMARY_SOURCE_FILTER_FETCH_LIMIT
 from mex.editor.exceptions import escalate_error
-from mex.editor.search.models import SearchPrimarySource, SearchResult
+from mex.editor.search.models import (
+    ReferenceFieldFilter,
+    ReferenceFieldIdentifierFilter,
+    SearchPrimarySource,
+    SearchResult,
+)
 from mex.editor.search.transform import transform_models_to_results
 from mex.editor.state import State
 from mex.editor.utils import resolve_editor_value
 
 if TYPE_CHECKING:
     from reflex.istate.data import RouterData
+
+
+def _build_dynamic_refresh_params(reference_field_filter: ReferenceFieldFilter) -> dict:
+    identifiers = [
+        x for x in reference_field_filter.identifiers if not x.validation_msg
+    ]
+    if not reference_field_filter.field or not identifiers:
+        return {"reference_field": None, "referenced_identifier": None}
+    return {
+        "reference_field": reference_field_filter.field,
+        "referenced_identifier": [x.value for x in identifiers],
+    }
+
+
+def _build_had_primary_source_refresh_params(
+    had_primary_sources: dict[str, SearchPrimarySource],
+) -> dict:
+    had_primary_source = [
+        identifier
+        for identifier, primary_source in had_primary_sources.items()
+        if primary_source.checked
+    ]
+    return {
+        "reference_field": "hadPrimarySource" if had_primary_source else None,
+        "referenced_identifier": had_primary_source,
+    }
 
 
 class SearchState(State):
@@ -28,9 +62,91 @@ class SearchState(State):
     total: int = 0
     query_string: str = ""
     entity_types: dict[str, bool] = {k.stemType: False for k in MERGED_MODEL_CLASSES}
+
+    reference_filter_strategy: Literal["had_primary_source", "dynamic"] = "dynamic"
     had_primary_sources: dict[str, SearchPrimarySource] = {}
     current_page: int = 1
+    reference_field_filter: ReferenceFieldFilter = ReferenceFieldFilter(
+        field="", identifiers=[]
+    )
     is_loading: bool = True
+
+    @rx.event
+    def set_reference_filter_field(self, value: str) -> None:
+        """Set the reference filter field used to filter references.
+
+        Args:
+            value (str): The field to use for reference filtering.
+        """
+        self.reference_field_filter.field = value
+
+    @rx.event
+    def set_reference_filter_strategy(
+        self, value: Literal["had_primary_source", "dynamic"]
+    ) -> None:
+        """Set the reference filter strategy to define the filter mechanism.
+
+        Args:
+            value: The strategy used for filtering.
+        """
+        self.reference_filter_strategy = value
+
+    @rx.event
+    def set_reference_field_filter_identifier(self, index: int, value: str) -> None:
+        """Set the reference value to filter for at a specific index.
+
+        Args:
+            index (int): Index of the identifier
+            value (str): Value of the identifier
+        """
+        self.reference_field_filter.identifiers[index].validation_msg = None
+        self.reference_field_filter.identifiers[index].value = value
+        try:
+            TypeAdapter(Identifier).validate_python(value)
+        except ValidationError as ve:
+            self.reference_field_filter.identifiers[index].validation_msg = "\n".join(
+                [error["msg"] for error in ve.errors()]
+            )
+
+    @rx.event
+    def remove_reference_field_filter_identifier(self, index: int) -> None:
+        """Remove the reference valueto filter for at a specific index.
+
+        Args:
+            index (int): Index of the identifier to remove.
+        """
+        self.reference_field_filter.identifiers.pop(index)
+
+    @rx.event
+    def add_reference_field_filter_identifier(self) -> None:
+        """Add a new empty identifier."""
+        self.reference_field_filter.identifiers.append(
+            ReferenceFieldIdentifierFilter(value="", validation_msg=None)
+        )
+        self.set_reference_field_filter_identifier(
+            len(self.reference_field_filter.identifiers) - 1, ""
+        )
+
+    @rx.var(cache=False)
+    def all_fields_for_entity_types(self) -> list[str]:
+        """Get all fields for the currently selected entity types filter.
+
+        Returns:
+            list[str]: The fields for the selected entity types.
+        """
+        selected_entity_types = [
+            item[0]
+            for item in list(filter(lambda item: item[1], self.entity_types.items()))
+        ]
+
+        if len(selected_entity_types) == 0:
+            selected_entity_types = [item[0] for item in self.entity_types.items()]
+
+        fields_by_type = [
+            REFERENCE_FIELDS_BY_CLASS_NAME[ensure_prefix(entity_type, "Extracted")]
+            for entity_type in selected_entity_types
+        ]
+        return sorted({f for fields in fields_by_type for f in fields})
 
     @rx.var(cache=False)
     def disable_previous_page(self) -> bool:
@@ -78,6 +194,28 @@ class SearchState(State):
         for primary_source_identifier in had_primary_source_params:
             self.had_primary_sources[primary_source_identifier].checked = True
 
+        reference_filter_strategy = router.page.params.get(
+            "referenceFilterStrategy", "dynamic"
+        )
+        self.reference_filter_strategy = (
+            reference_filter_strategy
+            if reference_filter_strategy in ["dynamic", "had_primary_source"]
+            else "dynamic"
+        )
+        ref_field_identifiers = router.page.params.get("referenceIdentifier", [])
+        ref_field_identifiers = (
+            ref_field_identifiers
+            if isinstance(ref_field_identifiers, list)
+            else [ref_field_identifiers]
+        )
+        self.reference_field_filter = ReferenceFieldFilter(
+            field=router.page.params.get("referenceField", ""),
+            identifiers=[
+                ReferenceFieldIdentifierFilter(value=x, validation_msg=None)
+                for x in ref_field_identifiers
+            ],
+        )
+
     @rx.event
     def push_search_params(self) -> Generator[EventSpec | None, None, None]:
         """Push a new browser history item with updated search parameters."""
@@ -86,10 +224,15 @@ class SearchState(State):
                 "q": self.query_string,
                 "page": self.current_page,
                 "entityType": [k for k, v in self.entity_types.items() if v],
+                "referenceFilterStrategy": self.reference_filter_strategy,
                 "hadPrimarySource": [
                     k for k, v in self.had_primary_sources.items() if v.checked
                 ],
-            },
+                "referenceField": self.reference_field_filter.field,
+                "referenceIdentifier": [
+                    x.value for x in self.reference_field_filter.identifiers
+                ],
+            }
         )
 
     @rx.event
@@ -156,11 +299,13 @@ class SearchState(State):
         entity_type = [
             ensure_prefix(k, "Merged") for k, v in self.entity_types.items() if v
         ]
-        had_primary_source = [
-            identifier
-            for identifier, primary_source in self.had_primary_sources.items()
-            if primary_source.checked
-        ]
+
+        filter_strategy_params = (
+            _build_dynamic_refresh_params(self.reference_field_filter)
+            if self.reference_filter_strategy == "dynamic"
+            else _build_had_primary_source_refresh_params(self.had_primary_sources)
+        )
+
         skip = DEFAULT_FETCH_LIMIT * (self.current_page - 1)
         self.is_loading = True
         yield None
@@ -168,10 +313,9 @@ class SearchState(State):
             response = connector.fetch_preview_items(
                 query_string=self.query_string,
                 entity_type=entity_type,
-                reference_field="hadPrimarySource" if had_primary_source else None,
-                referenced_identifier=had_primary_source,
                 skip=skip,
                 limit=DEFAULT_FETCH_LIMIT,
+                **filter_strategy_params,
             )
         except HTTPError as exc:
             self.is_loading = False
@@ -191,10 +335,11 @@ class SearchState(State):
     def get_available_primary_sources(self) -> Generator[EventSpec, None, None]:
         """Get all available primary sources."""
         connector = BackendApiConnector.get()
-        maximum_number_of_primary_sources = 100
         try:
             primary_sources_response = connector.fetch_preview_items(
-                entity_type=[MergedPrimarySource.__name__],
+                entity_type=[ensure_prefix(MergedPrimarySource.stemType, "Merged")],
+                skip=0,
+                limit=PRIMARY_SOURCE_FILTER_FETCH_LIMIT,
             )
         except HTTPError as exc:
             yield from escalate_error(
@@ -204,9 +349,9 @@ class SearchState(State):
             available_primary_sources = transform_models_to_results(
                 primary_sources_response.items
             )
-            if len(available_primary_sources) == maximum_number_of_primary_sources:
+            if len(available_primary_sources) == PRIMARY_SOURCE_FILTER_FETCH_LIMIT:
                 msg = (
-                    f"Cannot handle more than {maximum_number_of_primary_sources} "
+                    f"Cannot handle more than {PRIMARY_SOURCE_FILTER_FETCH_LIMIT} "
                     "primary sources."
                 )
                 raise MExError(msg)
@@ -224,3 +369,11 @@ class SearchState(State):
                     search_primary_sources, key=lambda source: source.title.lower()
                 )
             }
+
+
+full_refresh = [
+    SearchState.go_to_first_page,
+    SearchState.push_search_params,
+    SearchState.refresh,
+    SearchState.resolve_identifiers,
+]